% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data3d.R
\docType{data}
\name{data3d}
\alias{data3d}
\title{Three dimensional data set}
\format{
A frame (excluding the headings) with 156 lines of data and 5 variables:
\describe{
\item{K}{an integer ranging from 1 to 13, the size of an i.i.d. random sample from a 2 x 3 x 2 tensor matrix normal distribution}
\item{Id3}{an integer ranging from 1 to 2, the number of rows of the 3rd-order tensor normal distribution}
\item{Id4}{an integer ranging from 1 to 3, the number of columns of the 3rd-order tensor normal distribution}
\item{Id5}{an integer ranging from 1 to 2, the number of edges of the 3rd-order tensor normal distribution}
\item{value3d}{the sample data for the observed variable}
}
}
\usage{
data3d
}
\description{
An i.i.d. random sample of size 13 from a 2 x 3 x 2
tensor normal distribution, for a small numerical example of the use of the
functions mle3d_svc and lrt3d_svc from the \pkg{sEparaTe} package
}
\keyword{datasets}
